# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.2.1"
  description: "API made for the SHCITY project"
  title: SHCity API
  contact: 
    name: "Alexandre Dias"
    url: "https://www.linkedin.com/in/alexandredias95/"
    email: "alex.m.a.dias95@gmail.com" 
# during dev, should point to your local machine
host: 193.136.127.202:80
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /panels/add: 
    x-swagger-router-controller: panels
    post:
      summary: Add a Panel.
      tags:
              - panels
      consumes:
        - application/json
      operationId: panel_add
      parameters:
        - in: body
          name: Panel
          description: The Panel to add value.
          schema:
            $ref: "#/definitions/Panel"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"   
  /panels/panel_visitors_week_add: 
    x-swagger-router-controller: panels
    post:
      summary: Add a Panel value.
      tags:
              - panels
      consumes:
        - application/json
      operationId: panel_panel_visitors_week_add
      parameters:
        - in: body
          name: Panel
          description: The Panel to add value.
          schema:
            $ref: "#/definitions/Panel_week"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse" 
  /panels/panel_visitors_month_add: 
    x-swagger-router-controller: panels
    post:
      summary: Add a Panel value.
      tags:
              - panels
      consumes:
        - application/json
      operationId: panel_panel_visitors_month_add
      parameters:
        - in: body
          name: Panel
          description: The Panel to add value.
          schema:
            $ref: "#/definitions/Panel_month"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"  
  /panels/panel_visitors_year_add: 
    x-swagger-router-controller: panels
    post:
      summary: Add a Panel value.
      tags:
              - panels
      consumes:
        - application/json
      operationId: panel_panel_visitors_year_add
      parameters:
        - in: body
          name: Panel
          description: The Panel to add value.
          schema:
            $ref: "#/definitions/Panel_year"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"
...
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  Panel:
    properties:  
        username:
          type : string
        token:
          type : string
        panel_name:
          type : string
        description:
          type : string  
        location:
          type : string
  Panel_week:
    properties:  
        username:
          type : string
        token:
          type : string
        panel_name:
          type : string
        description:
          type : string  
        visitors:
          type : string
        week:
          type : string  
        month:
          type : string  
        year:
          type : string
  Panel_month:
    properties:  
        username:
          type : string
        token:
          type : string
        panel_name:
          type : string
        description:
          type : string  
        visitors:
          type : string
        month:
          type : string  
        year:
          type : string
  Panel_year:
    properties:  
        username:
          type : string
        token:
          type : string
        panel_name:
          type : string
        description:
          type : string  
        visitors:
          type : string
        year:
          type : string  
...